Removed query chaining as it doesn't make sense in the ValueType context.
For example, attempting to chain a ValueType will always return an immutable version of the query, which won't allow you to make further mutations to the returned copy of the query you are attempting to chain (if it was a ReferenceType it would work).
If you really want a mutable version to chain, you would have to declare a new variable with "var", or as the proper way in the ValueType context, start the initial query with "var" and make all mutations on that one query.

This isn't needed in Parse-Swift as the "where" of the query can be easily appended/mutated, therefore providing the the ability to "chain" anyways